[DEV] Move BaseObject to s/o/base.py
